src/pyproject.toml
[project]
name = "dr-lawyer-legal-intake-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=3.0.3",
    "langgraph-cli>=0.4.11",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "langgraph-cli[inmem]<=0.4.11",
]
